requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
google-generativeai==0.3.2
numpy==1.26.2
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3
gunicorn==21.2.0
python-dotenv==1.0.0